Add snapcraft.yaml defining the snap package and its components
The ldc2 package uses `classic` confinement to ensure that it can access
the resources of the host system it is installed on, such as gcc, libc,
and development libraries against which compiled code needs to link.

The commands exposed include not just `ldc2` itself but also `ldmd2` and
the `ldc-profdata` and `ldc-prune-cache` utilities.  These latter three
commands are for now callable only as `ldc2.ldmd2`, `ldc2.ldc-profdata`,
and `ldc2.ldc-prune-cache`, due to snap-packaging naming constraints on
publicly exposed commands.

Note that `ldc` itself is required as a build dependency, since the LDC
build process requires a working D compiler to compile the DMD frontend
(itself written in D).  On an Ubuntu 16.04 system, LDC 0.17.1 will be
used (the last LDC version to use the old C++ DMD frontend).

A couple of packages related to the build dependencies (libphobos2-ldc68
and libconfig9) are included as staged packages, since the classic snap
needs these available to link against.