Explicitly specify the LLVM version to use
It seems saner to be explicit about the LLVM version than to rely on a
distro-dependent version that may change.  Ubuntu 16.04 does not provide
a package for LLVM 3.9, so 3.8 is the most recent version that can be
used unless the snap builds LLVM itself.  This may be worth implementing
in future, but it would add heavily to the build time and so has been
avoided for now.