Update LDC version to v1.1.1
v1.1.1-beta1 proved a viable solution to the position-independent code
issue (cf. ldc-developers/ldc2.snap#15) so has been updated to the final
v1.1.1 release.

This patch updates the snap-package definition accordingly and re-grades
the snap once more as `stable`, re-enabling the possibility to upload to
the candidate and release channels.