Revert "Add MULTILIB support to the LDC build"
Enabling multilib for 32-bit builds currently results in build failure
due to a conflict between the multilib `_64` suffix and modules that
normally have `_64` in their name.

Although it might be possible to still enable multilib for 64-bit builds
only, this is put off for a later time in order to not block other new
features from being possible to roll out.

This reverts commit c2770f024b67b2ec90cce3cf551850bacd9ed575.