Run ctest after LDC build is complete
This patch adds a new `ctest-setup` part to install all dependencies for
the `ctest` collection of tests defined for LDC.  The somewhat misnamed
`install:` scriptlet (which actually runs post-install) is then used in
the `ldc` part to invoke `ctest` itself, in verbose mode so that all the
tests will be clearly described in the build log.

This should ensure that the packaged compiler is guaranteed to have at
least passed the standard battery of tests on the build platform.

Addresses part of https://github.com/ldc-developers/ldc2.snap/issues/29.