Disable lit-tests entirely
The previous patch failed to work adequately, as it looks like the lit
tests are building from source files in `parts/ldc/src` instead of in
`parts/ldc/build`.

This feels like an uncomfortable conflation of the different parts of
the source tree, so in order to move forward, it seems simpler to just
disable the lit-tests entirely for now.  We can come back to them at a
later date, either once the issues with them have been fixed, or with
some later tweak that better manages exclusion of the problem tests.

The `ctest-setup` part, which installs python dependencies required by
the lit-tests, has been left untouched in order to ease the transition
back to enabling these tests.