Re-enable MULTILIB support for the LDC build
This should ensure that both 32- and 64-bit libraries are available in
the snap package, whatever the target architecture of the build.  With
various issues addressed since the last time we tried this, it seems
worth repeating the attempt.

`ldc2.conf` has been updated to handle all possible expected library
paths: `lib` for native, `lib32` for 32-bit libraries in a 64-bit build,
and `lib64` for 64-bit libraries in a 32-bit build.