Add CircleCI build config
This patch adds a basic CircleCI config file in line with the suggested
setup provided by
https://tutorials.ubuntu.com/tutorial/continuous-snap-delivery-from-circle-ci

This version implements only the build stage, and does not attempt to
push releases to the snap store; this will still be taken care of for
now by the Launchpad build system.  Its intended usage is to test PRs
and the current state of release branches.

`--no-parallel-builds` is used to avoid excessive resource consumption
on CircleCI.