Add basic CI support for Azure Pipelines
Since snapd is available in the Azure Ubuntu 16.04 image, we might as
well keep it simple, install snapcraft as a snap package and run it.

This should pretty much match what happens on Launchpad, apart from the
lack of LXD container.

Given the build times experienced in Launchpad, the timeout has been set
to 0, meaning it will always be the maximum allowed by Azure Pipelines
(currently 360 minutes).