Explicitly manage license file to avoid snapcraft 3 deprecation
Placing `license.txt` in the `snap` folder was always a rather flaky way
to get it in the package, and snapcraft 3 does not like arbitrary files
being placed here.  This patch therefore creates a new part which places
the license file in the `$SNAP/doc` directory of the generated package.