Replace ldc-bootstrap part with build-snaps
LDC v1.12.0+ supports LLVM 7, but v0.17.6 does not.  This means that to
build more recent LDC releases with the latest LLVM we will either need
to build multiple LLVMs (!) or change the bootstrap compiler.

With build-snaps now a robust feature, this seems a good moment to tweak
the snap definition to use it, so that from here on out each release of
the snap package will be built with the previous compiler version.