Update to LLVM 9.0.1
This update involves switching repo to LDC's direct fork of the LLVM
upstream:
https://github.com/ldc-developers/llvm-project/tree/ldc-v9.0.1

Thanks to the absence of a base-directory-level `CMakeLists.txt` we have
to add a manual `override-build` to run `cmake` on the `llvm` directory.
`ninja` is used to match the upstream LDC build system.

CMake build flags have also been updated to bring them in line with the
choices in the LDC project CI.  Note that this includes updates to the
supported target architectures.