Add a base specifier to snapcraft.yaml
This should open up a number of new features, including the opportunity
to specify license details in snap package metadata, and other snapcraft
3 features.

See https://snapcraft.io/docs/base-snaps for details on base snaps.  The
`core` base is chosen so as to ensure compatibility with Ubuntu 16.04.

One impact of this is that by default the `snapcraft` command will build
using `multipass` (and `snapcraft cleanbuild` will no longer run).  This
unfortunately does not play nicely with Azure Pipelines, so the build is
run with the `--destructive-mode` option to build directly inside the CI
environment.