Work around Azure Piplines build problems by chown'ing root directory
Using `chown` to set `root` as the owner of `/` was suggested on the
snapcraft forums, and appears to work for others:
https://forum.snapcraft.io/t/permissions-problem-using-snapcraft-in-azure-pipelines/13258/14

Unfortunately this exposes a nasty regression in the DMD testsuite: the
`ddocYear` test-case hard-encodes an expected year, when it should be a
placeholder that gets replaced with the year when the build takes place.
We can work around this by using `sed` to replace the hardcoded year.

It will be necessary to update this workaround for LDC 1.20.x (when the
hardcoded year switches from 2019 to 2020), and it should be possible to
remove the workaround entirely with LDC 1.21, thanks to an upstream fix
that should land in the next DMD frontend release:
https://github.com/dlang/dmd/pull/10802