Add license metadata to snapcraft.yaml
This will allow the package to display license info in the snap store
and on the command-line `snap info` results.

The license spec is provided in SPDX format <https://spdx.org> with the
following components:

  * BSD-3-Clause (for LDC itself)

  * BSL-1.0 (DMD/druntime/phobos)

  * Zlib (zlib, statically compiled into phobos)

  * Apache-2.0 AND NCSA (LLVM)

  * Artistic-1.0 OR GPL-2.0+ (GDC artefacts)

Note that snapcraft and snapd only support SPDX 2.x license identifiers,
so we must use `GPL-2.0+` instead of the more recent `GPL-2.0-or-later`.
See:
https://forum.snapcraft.io/t/keeping-up-with-spdx-license-identifier-updates/7095