Update ldc2.conf to better match upstream 1.18.0 binary package
This brings the config file in line with upstream standards.  It should
be backwards compatible, while adding support for some extra settings,
such as WebAssembly builds.

The only differences from upstream are the import paths, which are set
to `include/d` instead of `import` (accepting the default install path
of the build system), and the library paths, which are set to `lib64`
and `lib32` instead of `lib` and `lib32` (making the paths consistent
for both amd64 and i386 package builds).