Simplify or remove override-build sections
`snapcraft` now actively supports using the ninja cmake backend, so in
general we do not need to `override-build` to use this.  This allows us
to go back to regular `configflags` settings for the `ldc-bootstrap` and
`llvm` parts, although we have to add `source-subdir` settings to `llvm`
because its `CMakeLists.txt` file is not in the source root.

For the `ldc` part we can move all the cmake config flags and simplify
the `override-build` to just run `snapcraftctl build` followed by the
search-and-replace dmd-testsuite workaround, and the call to `ctest`.

We also need to make one small tweak to the `ldc` part `configflags`:
when setting `DLIB_SUFFIX`, we need to use `$$` to escape the `$` for
ninja, and drop the surrounding quotes in order to ensure the scripted
segment is passed to cmake correctly by snapcraft.