Use clang(++) as the C(++) compiler to build LLVM and LDC
Using clang as our compiler both brings the snap package in line with
upstream LDC build practices and also opens the door to using LTO when
building LDC.

This patch takes advantage of the new `package-repositories` feature of
snapcraft 4 to add the LLVM apt repo, which we can use to install the
required clang(++) versions.  YAML anchors/aliases are used to store the
`c-compiler` and `cxx-compiler` choices so that these can be re-used in
the 3 different parts.  These choices need to be reverted to `gcc`/`g++`
when running the test suite in order to match upstream.

Since this is an experimental feature, we also need to update CI to use
the `--enable-experimental-package-repositories` snapcraft flag.

Finally, we need to drop the `-static-libstdc++` flag to avoid LLVM link
errors related to `-fPIC`.